c: player component structure refactoring
